Callback Handler Wrapper
Created a new class specialized for callback handling. I want to be able
to run callbacks on specific threads but having to surround it with
if-else statements each time hurts my OCD.

I ended up writing this wrapper class to handle it in one line of code
instead. It's much neater this way.