Rewrote CallbackHandler & support for safe callbacks
The CallbackHandler was initially written to provide clean code. However, it was just hiding it and eventually got too ugly. It's has now been rewritten to use reflection and thereby keeping things short and sweet.

Also, you now have the ability to safely handle callbacks (be it via singleton or per request). This is extremely useful in certain occasions but please do not overuse it as it may cause bugs without you knowing due to it catching them.