Completely revamped callbacks
I admit, my previous approach to calls was very flawed. The use of reflection in order to make calls look prettier was an even bigger mistake.

Callbacks now have one function instead of 3. The preload function was removed completely and the loaded/fail functions have been merged into one. If the request is successful, exception will be null and the loaded items will be non-null. The reverse is true if the request failed.

I've also removed the "safe" property. Silently catching exceptions thrown by callback code is bad practice.

Additionally, I've introduced the option to select a reference type and a convenience async function with callback included. Callbacks will continue to be stored as WeakReferences by default, but you'll be able to store them as strong references if you so wish. This is useful when you use anonymous callback instances, for example.